agent_service/requirements.txt
langchain
langgraph
langchain-community
python-dotenv
requests
langchain-openai
fastapi
uvicorn
